Shutdown's commit.

+++ shutdown.py
@@ -1,10 +1,25 @@
+# !bin/bash
 from os import system
+from sys import platform
+
 
 class Shutdown:
-    def __init__() -> None:
-        shutdown_command = 'shutdown -s -f -t 0'
-        system(shutdown_command)
+    """A class for shutdown the computer."""
 
-if __name__ == '__main__':
+    def __init__(self) -> None:
+        shutd = Shutdown.get_os_command()
+        if shutd != "Invalid System!":
+            system(shutd)
+
+    @staticmethod
+    def get_os_command() -> str:
+        if platform in ("linux", "linux2", "darwin"):
+            return "shutdown now"
+        elif platform in ("win32", "cygwin", "msys"):
+            return "shutdown -s -f -t 0"
+        else:
+            return "Invalid System!"
+
+
+if __name__ == "__main__":
     s = Shutdown()
-
